commited New changes

+++ app.py
@@ -1,6 +1,7 @@
 from flask import Flask, jsonify, Response
 import json
 import instaloader
+import os
 import re
 
 app = Flask(__name__)
@@ -16,13 +17,32 @@ def create_instaloader_instance():
     user_agent = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.3"
     L.context._session.headers["User-Agent"] = user_agent
 
+    # try:
+    #     with open(SESSION_FILE, 'rb') as session_file:
+    #         L.context.load_session_from_file(INSTAGRAM_USERNAME,session_file)
+    # except instaloader.exceptions.QueryReturnedNotFoundException:
+    #     L.context.login(INSTAGRAM_USERNAME, INSTAGRAM_PASSWORD)
+    #     with open(SESSION_FILE, 'wb') as session_file:
+    #         L.context.save_session_to_file(session_file)
+
     try:
-        with open(SESSION_FILE, 'rb') as session_file:
-            L.context.load_session_from_file(INSTAGRAM_USERNAME,session_file)
-    except instaloader.exceptions.QueryReturnedNotFoundException:
-        L.context.login(INSTAGRAM_USERNAME, INSTAGRAM_PASSWORD)
-        with open(SESSION_FILE, 'wb') as session_file:
-            L.context.save_session_to_file(session_file)
+        if os.path.exists(SESSION_FILE):
+            with open(SESSION_FILE, 'rb') as session_file:
+                L.context.load_session_from_file(INSTAGRAM_USERNAME, session_file)
+            
+            if not L.context.is_logged_in:
+                L.context.login(INSTAGRAM_USERNAME, INSTAGRAM_PASSWORD)
+                with open(SESSION_FILE, 'wb') as session_file:
+                    L.context.save_session_to_file(session_file)
+        else:
+            L.context.login(INSTAGRAM_USERNAME, INSTAGRAM_PASSWORD)
+            with open(SESSION_FILE, 'wb') as session_file:
+                L.context.save_session_to_file(session_file)
+    except instaloader.exceptions.QueryReturnedNotFoundException as e:
+        print(f"QueryNotFoundException: {e}")
+    except instaloader.exceptions.InstaloaderException as e:
+        print(f"InstaloaderException: {e}")
+
 
     proxies = {
         'http': 'socks5://yoqytafd-6:2dng483b96qx@p.webshare.io:80',
